Setup main logging under logger rpi.*
This commit setups logging under Python's logging library. More specifically, any
time Python's logging is used, the message will be outputed to the
console (stdout) and to a file under logs/main. The file is named
"log_(current date and time).log" and a new file is created for each run
of the program. Furthermore, the logs follow this format:

	(date + time) - (logger's name) - (logging level) - (message)