Use logging library to log data from sensors
The Python's logging library is now used for logging data from sensors.
The logs location for any sensor data is in logs/sensor/{name of sensor},
with the file being a csv file. The abstract class SensorLogging handles
all the initialization for each sensor type, so it would be best if all
sensors' classes are a subclass of SensorLogging.

A filter to the root logger as been added for sensor logging. This means
that any logger name starting with "sensorlog" will not have the global
logging rules applied.