Add status display to GUI
Added a new window with a status indicator. This indicator is by default
green, and will change color if a message is logged with a level of
WARNING or ERROR. Furthermore, for these levels, the indicator will
flash to be more apparent to the user. The indicator can be cleared back
to green using a button.