Fix laptop not creating sensor log files
Add checks when the logging receiver receives a log with the logger name
starting with 'sensorlog'. If this is the case, it will then check if it
has a handler attached, and create it if this is not the case.

The end result is that the log files follow the same format as the RPI,
where there is a seperate file for each sensor.