Add log message to buffer if it cannot be sent
By default, logging.handlers.SocketHandler will silently drop the
message if it cannot be sent. Because this could be unwanted, this
commits adds a BufferedSocketHandler that will save the message if it
could not be sent. When a connection is made with the laptop, all the
unsent messages will be sent in the correct order.