Use config file for IP addresses
Use a file called 'config.ini' to save the IP and port for the RPI and
laptop. If the file doesn't exist, it will be created using the default
values.

Please note 'config.ini' is ignored by git. Because of this, different
computers can have different config files without being bothered by git.