Add option to clear status automatically in GUI if the issue is solved
When a log message is created, it has the option of specifying a
'errorID' value in the argument 'extra'. This enables the status frame
to track if an error has been cleared and automatically clear the
warning or error.

More specifically, if a log message of level WARNING or ERROR has a
'errorID' value, status frame will warn the user as usual and save this
errorID. If a message with a level of INFO arrives after with the same
errorID, the status frame will match this with the stored errorID and
automatically clear the error. Furthermore, if a message clears
automatically the status, it will be displayed in green to the user to
help see that the issue has been indeed solved.