Implement a loop that continously ping the RPI
While there is a constant communication between the RPI and laptop for
the logs, messages to control the motors and LEDs will be more sparse.
To prevent surprises when a control message is sent, the laptop will
regularly ping the RPI to make sure communication is fine. If there is a
problem, user is alerted.