Add check if the sensor is not connected
The register PROD_ID contains a known value from the sensor. This adds a
check to see if the value is indeed was is expected. If the value
returned is not as expected, it may mean that the sensor is disconnected
or that a different kind of sensor is connected.